requirements.txt
Flask==3.0.0
cryptography==41.0.7
argon2-cffi==23.1.0
python-dotenv==1.0.0
